- New Dazzling Gleam animation.
Note: the previous animation showed sprites over the player and then over the enemy. The engine seems to have trouble with that when the enemy uses the animation. I wasn't able to find how to program it as the intended one, so I made a new one using a background effect over the player and a single location for the other animation sprites over the enemy.